@@ -1,60 +1,67 @@
 {
-  "name": "dokku-overview",
-  "version": "1.0.0",
-  "description": "",
+  "name": "spectacle-boilerplate",
+  "version": "1.0.1",
+  "description": "ReactJS Powered Presentation Framework",
   "main": "lib/index.js",
   "scripts": {
     "clean": "rimraf dist",
     "build": "cross-env NODE_ENV=production webpack --config webpack.config.production.js",
     "lint": "eslint --ext .js,.jsx .",
     "deploy": "npm run clean & npm run build && surge -p .",
-    "development": "cross-env NODE_ENV=development node server.js",
-    "start": "NODE_ENV=development node server.js"
+    "export": "spectacle-renderer --delay 3000",
+    "start": "cross-env NODE_ENV=development node server.js"
   },
-  "author": "Maksym Popenko <popenkomaksim@gmail.com>",
+  "author": "",
   "license": "MIT",
   "dependencies": {
-    "normalize.css": "3.0.3",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-loadable": "^3.3.1",
-    "react-typist": "^1.1.1",
-    "spectacle": "^2.1.0",
-    "spectacle-code-slide": "^0.4.0",
-    "spectacle-terminal": "^0.3.0",
-    "spectacle-theme-nova": "^1.3.0"
+    "normalize.css": "7.0.0",
+    "prop-types": "15.6.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "spectacle": "^4.0.0",
+    "spectacle-renderer": "^0.0.3",
+    "spectacle-terminal": "0.5.0"
   },
   "devDependencies": {
-    "autoprefixer-core": "^6.0.1",
-    "babel-core": "^6.4.0",
-    "babel-eslint": "^5.0.0-beta6",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-react-transform": "^2.0.2",
-    "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-0": "^6.5.0",
-    "cross-env": "^1.0.7",
-    "css-loader": "^0.25.0",
-    "eslint": "^1.8.0",
-    "eslint-config-defaults": "^7.1.1",
-    "eslint-plugin-filenames": "^0.1.2",
-    "eslint-plugin-react": "^3.6.3",
-    "express": "^4.14.0",
-    "file-loader": "^0.9.0",
-    "html-loader": "^0.4.4",
-    "is-buffer": "^1.1.4",
-    "markdown-loader": "^0.1.7",
-    "node-libs-browser": "^0.5.3",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.1",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-react-transform": "^3.0.0",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-react-constant-elements": "^6.23.0",
+    "babel-plugin-transform-react-inline-elements": "^6.22.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.10",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-runtime": "^6.26.0",
+    "cross-env": "^5.1.0",
+    "css-loader": "^0.28.7",
+    "eslint": "^4.9.0",
+    "eslint-config-formidable": "^3.0.0",
+    "eslint-plugin-filenames": "^1.2.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-react": "^7.4.0",
+    "express": "^4.16.2",
+    "file-loader": "^1.1.5",
+    "html-loader": "^0.5.1",
+    "is-buffer": "^1.1.5",
+    "markdown-loader": "^2.0.1",
+    "node-libs-browser": "^2.0.0",
     "raw-loader": "^0.5.1",
-    "react-transform-hmr": "^1.0.4",
-    "rimraf": "^2.5.4",
-    "style-loader": "^0.13.1",
+    "react-hot-loader": "^3.1.1",
+    "react-transform-catch-errors": "^1.0.2",
+    "redbox-react": "^1.5.0",
+    "rimraf": "^2.6.2",
+    "style-loader": "^0.19.0",
     "surge": "latest",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.2",
-    "webpack-dev-middleware": "^1.8.3",
-    "webpack-hot-middleware": "^2.12.2"
+    "url-loader": "^0.6.2",
+    "webpack": "3.8.1",
+    "webpack-dev-middleware": "^1.12.0",
+    "webpack-dev-server": "^2.9.3",
+    "webpack-hot-middleware": "^2.20.0"
   }
 }
