@@ -8,10 +8,9 @@ var config = require("./webpack.config");
 var app = express();
 var compiler = webpack(config);
 
-var serverPort = process.env.PORT || 3000;
+var serverPort = process.env.PORT || 5555;
 
 app.use(require("webpack-dev-middleware")(compiler, {
-  noInfo: true,
   publicPath: config.output.publicPath
 }));
 
