@@ -14,7 +14,6 @@ module.exports = {
     publicPath: "/dist/"
   },
   plugins: [
-    new webpack.optimize.OccurenceOrderPlugin(),
     new webpack.DefinePlugin({
       "process.env": {
         "NODE_ENV": JSON.stringify("production")
@@ -33,10 +32,7 @@ module.exports = {
     }, {
       test: /\.(js|jsx)$/,
       exclude: /node_modules/,
-      loader: "babel-loader",
-      query: {
-        presets: ['es2015', 'react']
-      }
+      loader: "babel-loader"
     }, {
       test: /\.css$/,
       loader: "style-loader!css-loader"
@@ -45,7 +41,7 @@ module.exports = {
       loader: "url-loader?limit=8192"
     }, {
       test: /\.svg$/,
-      loader: "url?limit=10000&mimetype=image/svg+xml"
+      loader: "url-loader?limit=10000&mimetype=image/svg+xml"
     }]
   }
 };
